Move logic to link preview dialog
