+++ app/src/main/java/org/wikipedia/page/linkpreview/LinkPreviewViewState.kt
@@ -9,4 +9,5 @@ sealed class LinkPreviewViewState {
     data class Error(val throwable: Throwable) : LinkPreviewViewState()
     data class Content(val data: PageSummary) : LinkPreviewViewState()
     data class Gallery(val data: List<MwQueryPage>) : LinkPreviewViewState()
+    data class Watch(val isWatch: Boolean) : LinkPreviewViewState()
 }
