+++ app/src/main/java/org/wikipedia/places/PlacesFragmentViewModel.kt
@@ -11,7 +11,6 @@ import com.mapbox.mapboxsdk.plugins.annotation.Symbol
 import kotlinx.coroutines.CoroutineExceptionHandler
 import kotlinx.coroutines.launch
 import org.wikipedia.Constants
-import org.wikipedia.analytics.eventplatform.WatchlistAnalyticsHelper
 import org.wikipedia.dataclient.ServiceFactory
 import org.wikipedia.dataclient.WikiSite
 import org.wikipedia.extensions.parcelable
@@ -19,8 +18,6 @@ import org.wikipedia.page.PageTitle
 import org.wikipedia.settings.Prefs
 import org.wikipedia.util.ImageUrlUtil
 import org.wikipedia.util.Resource
-import org.wikipedia.util.log.L
-import org.wikipedia.watchlist.WatchlistExpiry
 
 class PlacesFragmentViewModel(bundle: Bundle) : ViewModel() {
 
@@ -28,13 +25,6 @@ class PlacesFragmentViewModel(bundle: Bundle) : ViewModel() {
     var location: Location? = bundle.parcelable(PlacesActivity.EXTRA_LOCATION)
     var pageTitle: PageTitle? = bundle.parcelable(Constants.ARG_TITLE)
 
-    var watchlistExpiryChanged = false
-    var isWatched = false
-    var hasWatchlistExpiry = false
-    var lastWatchExpiry = WatchlistExpiry.NEVER
-    var currentMarkerPageTitle: PageTitle? = null
-
-    val watchStatus = MutableLiveData<Resource<Boolean>>()
     val nearbyPages = MutableLiveData<Resource<List<NearbyPage>>>()
 
     fun fetchNearbyPages(latitude: Double, longitude: Double, radius: Int, maxResults: Int) {
@@ -53,38 +43,6 @@ class PlacesFragmentViewModel(bundle: Bundle) : ViewModel() {
                 }
 
             nearbyPages.postValue(Resource.Success(pages))
-        }
-    }
-
-    fun watchOrUnwatch(expiry: WatchlistExpiry, unwatch: Boolean) {
-        currentMarkerPageTitle?.let { title ->
-            if (isWatched) {
-                WatchlistAnalyticsHelper.logRemovedFromWatchlist(title)
-            } else {
-                WatchlistAnalyticsHelper.logAddedToWatchlist(title)
-            }
-            viewModelScope.launch(CoroutineExceptionHandler { _, throwable ->
-                L.w("Failed to fetch watch status.", throwable)
-            }) {
-                val token = ServiceFactory.get(title.wikiSite).getWatchToken().query?.watchToken()
-                val response = ServiceFactory.get(title.wikiSite)
-                    .watch(if (unwatch) 1 else null, null, title.prefixedText, expiry.expiry, token!!)
-
-                lastWatchExpiry = expiry
-                if (watchlistExpiryChanged && unwatch) {
-                    watchlistExpiryChanged = false
-                }
-                if (unwatch) {
-                    WatchlistAnalyticsHelper.logRemovedFromWatchlistSuccess(title)
-                } else {
-                    WatchlistAnalyticsHelper.logAddedToWatchlistSuccess(title)
-                }
-                response.getFirst()?.let {
-                    isWatched = it.watched
-                    hasWatchlistExpiry = lastWatchExpiry != WatchlistExpiry.NEVER
-                    watchStatus.postValue(Resource.Success(isWatched))
-                }
-            }
         }
     }
